Works but has a time limit
There is now a time limit for how long a turn can be for both the player and the Enemy AI. Which can be useful, but I want people to take their time if they want. Try to redo some classes and try to implement the Command Pattern. So the Ai can be easier to invoke and the Cards. Hopefully the Turn Manger will be easier to handle.